Lay groundwork for posting survey to API
Bring in redux-thunk to allow dispatching functions rather than action
objects. This allows us to receive dispatch in a callback so that we
can dispatch another action when the request completes/fails.

For HTTP transport, use a Fetch[1] polyfill (polyfill needed since
Fetch isn't implemented in browsers yet). Of course, as the fetch
library is implemented as a polyfill, it requires some *magic*
incantations[3] to get it to work properly with webpack:

    plugins: [
      new webpack.ProvidePlugin({
        'fetch': 'imports?this=>global!exports?global.fetch!whatwg-fetch'
      })
    ]

'that?snot-a=>cryptic!string.at@all'...

[1]: https://fetch.spec.whatwg.org/#concept-network-error
[2]: https://github.com/github/fetch
[3]: http://mts.io/2015/04/08/webpack-shims-polyfills/